chore: move tests into separate dir
