+++ program-libs/zero-copy-derive/src/tests/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(test)]
+mod from_impl;
+#[cfg(test)]
+mod utils;
+#[cfg(test)]
+mod z_struct;
